% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_GBIF_locs.R
\name{map_GBIF_locs}
\alias{map_GBIF_locs}
\title{Map the locations of the species using GBIF data}
\usage{
map_GBIF_locs(Species)
}
\arguments{
\item{Species}{The species name e.g. "Python bivittatus"}
}
\value{
ggplot map of species location records (can include non-native records)
}
\description{
Map the locations of the species using GBIF data
}
\examples{
map_GBIF_locs(Species="Morelia amethistina")

}
