% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_python_trade.R
\name{plot_python_trade_network}
\alias{plot_python_trade_network}
\title{Plot python trade network}
\usage{
plot_python_trade_network(Taxon, importer = "NO", min_year = 2009)
}
\arguments{
\item{Taxon}{Species name}

\item{importer}{Country of import (defaults to "NO" - Norway)}

\item{min_year}{A year filter default is >2009 (so start year is 2010)}
}
\value{
Visnetwork graph of origin, export and import of live and derived products of pythons
}
\description{
Plot python trade network
}
\examples{
plot_python_trade_network(Taxon="Python bivittatus")
plot_python_trade_network(Taxon="Morelia bredli")
}
