% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Purpose_codes.R
\docType{data}
\name{Purpose_codes}
\alias{Purpose_codes}
\title{Purpose_codes}
\format{
A data frame with 12 rows and 2 variables:
\describe{
  \item{\code{Codes}}{character Single letter code}
  \item{\code{description}}{character Decription of the codes meaning}
}
}
\usage{
Purpose_codes
}
\description{
Codes and description for the Purpose listed in the CITES Trade database
}
\details{
Codes to identify the purposes described in the CITES Trade Database
}
\keyword{datasets}
