% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_codes.R
\docType{data}
\name{Source_codes}
\alias{Source_codes}
\title{Source_codes}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{\code{Codes}}{character Single letter code}
  \item{\code{description}}{character Decription of the codes meaning}
}
}
\usage{
Source_codes
}
\description{
Codes and description for the Sources listed in the CITES Trade database
}
\details{
Codes to identify the source described in the CITES Trade Database
}
\keyword{datasets}
